Make JsonLdSerializer injectable
DI requires public types and constructors. Rather than just make everything public, I moved the the internal pivot classes to be internal within TypeConverter, which is the only place they're ever used. They can still be replaced by DI provided implementations, via the long constructor.
