+++ Source/ActivityPub.Types/Conversion/Converters/ListableConverter.cs
@@ -11,8 +11,9 @@ namespace ActivityPub.Types.Conversion.Converters;
 ///     Converts types that can be a list of elements, or a single elements.
 ///     Essentially: T | T[]
 /// </summary>
-internal class ListableConverter : JsonConverterFactory
+public class ListableConverter : JsonConverterFactory
 {
+    /// <inheritdoc />
     public override bool CanConvert(Type type) =>
         // We only convert concrete types deriving from ICollection<T>
         type.IsAssignableToGenericType(typeof(ICollection<>))
@@ -23,6 +24,7 @@ internal class ListableConverter : JsonConverterFactory
         && !type.IsAssignableToGenericType(typeof(IDictionary<,>));
 
     // Pivot the type into correct converter
+    /// <inheritdoc />
     public override JsonConverter? CreateConverter(Type collectionType, JsonSerializerOptions options)
     {
         var itemType = collectionType.GetGenericArgumentsFor(typeof(IReadOnlyCollection<>))[0];
