+++ Test/ActivityPub.Types.Tests/Util/Fixtures/JsonLdSerializerFixture.cs
@@ -5,7 +5,6 @@ using ActivityPub.Types.Conversion;
 using ActivityPub.Types.Conversion.Converters;
 using ActivityPub.Types.Conversion.Overrides;
 using ActivityPub.Types.Internal;
-using ActivityPub.Types.Internal.Pivots;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Options;
 
@@ -30,9 +29,9 @@ public sealed class JsonLdSerializerFixture
 
         ConversionOptions = new ConversionOptions();
         var conversionOptions = Options.Create(ConversionOptions);
-        var namelessEntityPivot = new NamelessEntityPivot();
-        var anonymousEntityPivot = new AnonymousEntityPivot();
-        var customConvertedEntityPivot = new CustomConvertedEntityPivot();
+        var namelessEntityPivot = new TypeMapConverter.NamelessEntityPivot();
+        var anonymousEntityPivot = new TypeMapConverter.AnonymousEntityPivot();
+        var customConvertedEntityPivot = new TypeMapConverter.CustomConvertedEntityPivot();
         var typeMapConverter = new TypeMapConverter(ASTypeInfoCache, conversionOptions, anonymousEntityPivot, namelessEntityPivot, customConvertedEntityPivot);
 
         var asTypeConverter = new ASTypeConverter();
